[Silver III] Title: 주유소, Time: 320 ms, Memory: 40604 KB, Score: 100 point -BaekjoonHub

+++ 백준/Silver/13305. 주유소/주유소.java
@@ -12,18 +12,20 @@ public class Main {
         for (int i = 0; i < cityDistanceList.length-1; i++) {
             cityDistanceList[i] = Integer.parseInt(st.nextToken());
         }
+
         st = new StringTokenizer(br.readLine());
         for (int i = 0; i < cityPayList.length; i++) {
             cityPayList[i] = Integer.parseInt(st.nextToken());
         }
+
         //계산
         int nextPos = 0;
-        int answer = 0;
+        long answer = 0;
         for (int p = 0; p < cityPayList.length; p++) {
             boolean isFlag = false;
             while (nextPos<cityTotal&&cityPayList[p]<=cityPayList[nextPos]){
                 isFlag = true;
-                answer += cityPayList[p]*cityDistanceList[nextPos];
+                answer += (long) cityPayList[p] *cityDistanceList[nextPos];
                 nextPos++;
             }
             if (isFlag) p = nextPos-1;
